Read server host, port from conf file. Added Route file for loading coroutine defs.
