+++ commandServer/command_server.conf
@@ -1,0 +1,3 @@
+[commandServer]
+host = 0.0.0.0
+port = 8000
